Additional modifications to tests were necessary.
