+++ src/son/schema/tests/son-schema/vnfd-schema.yml
